updated project - category

+++ src/main/java/com/gul/product/service/persistance/CategoryDao.java
@@ -14,7 +14,7 @@ public class CategoryDao extends AbstractDAO<Category> {
 		return persist(category);
 	}
 	
-	public Category findById(Integer id) {
+	public Category findById(Long id) {
 		return get(id);
 	}
